추천기능 백엔드 기능 구현

+++ V-Search/src/main/java/com/mysite/sbb/recommend/RecommendRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 public interface RecommendRepository extends JpaRepository<Recommend, Long> {
 	@Modifying
     @Query(value = "INSERT INTO recommend(VIDEO_NO, USER_NO) VALUES(:video_no, :user_no)", nativeQuery = true)
-    int recommend(Long video_no, int user_no);
+    int recommend(Long video_no, Long user_no);
 
     @Modifying
     @Query(value = "DELETE FROM recommend WHERE VIDEO_NO = :video_no AND USER_NO = :user_no", nativeQuery = true)
-    int cancelRecommend(Long video_no, int user_no);
+    int cancelRecommend(Long video_no, Long user_no);
 }
