+++ V-Search/src/main/java/com/mysite/sbb/recommend/RecommendService.java
@@ -11,12 +11,12 @@ public class RecommendService {
 	private final RecommendRepository recommendRepository;
 
 	@Transactional
-    public void recommend(Long video_no, int user_no) {
+    public void recommend(Long video_no, Long user_no) {
         recommendRepository.recommend(video_no, user_no);
     }
 
     @Transactional
-    public void cancelRecommend(Long video_no, int user_no) {
+    public void cancelRecommend(Long video_no, Long user_no) {
         recommendRepository.cancelRecommend(video_no, user_no);
     }
 }
